refactor: 전역적으로 쓰이는 컴포넌트들을 (home)경로가 아닌 그 위에 있도록 수정
