+++ src/app/(beforeLogin)/signUp/page.tsx
@@ -1,5 +1,5 @@
-import Input from '@/app/(home)/_component/ui/Input';
-import Button from '@/app/(home)/_component/ui/Button';
+import Input from '@/app/_component/ui/Input';
+import Button from '@/app/_component/ui/Button';
 
 export default function SignUp() {
   return (
